+++ src/Jukebox/Resample.cpp
@@ -3,39 +3,39 @@
 namespace Dynamo {
     void resample_signal(WaveSample *src,
                          WaveSample *dst,
-                         float time_offset,
-                         float src_length,
-                         float src_rate,
-                         float dst_rate) {
-        float factor = dst_rate / src_rate;
-        float scale = std::min(1.0f, factor);
-        float time_increment = 1.0 / factor;
+                         double time_offset,
+                         double src_length,
+                         double src_rate,
+                         double dst_rate) {
+        double factor = dst_rate / src_rate;
+        double scale = std::min(1.0, factor);
+        double time_increment = 1.0 / factor;
 
         unsigned dst_length = src_length * factor;
         unsigned filter_step = scale * FILTER_PRECISION;
 
-        float time = time_offset;
-        float time_end = src_length + time_offset;
+        double time = time_offset;
+        double time_end = src_length + time_offset;
 
         for (unsigned dst_f = 0; dst_f < dst_length; dst_f++) {
             // Target sample
-            float dst_sample = 0;
+            double dst_sample = 0;
 
             // Get the source frame
             unsigned src_f = time;
-            float P = scale * (time - src_f);
+            double P = scale * (time - src_f);
 
             // Calculate filter offset and interpolation factor
-            float P_frac = P * FILTER_PRECISION;
+            double P_frac = P * FILTER_PRECISION;
             unsigned l = P_frac;
-            float interp = P_frac - l;
+            double interp = P_frac - l;
 
             // Left wing
             for (int i = 0; i < FILTER_HALF_LENGTH; i++) {
                 int s_i = src_f - i;
                 int h_i = l + i * filter_step;
                 if (s_i < 0 || h_i >= FILTER_HALF_LENGTH) break;
-                float weight = FILTER_RWING[h_i] + interp * FILTER_DIFFS[h_i];
+                double weight = FILTER_RWING[h_i] + interp * FILTER_DIFFS[h_i];
                 dst_sample += src[s_i] * weight;
             }
 
@@ -52,7 +52,7 @@ namespace Dynamo {
                 int s_i = src_f + i + 1;
                 int h_i = l + i * filter_step;
                 if (s_i >= time_end || h_i >= FILTER_HALF_LENGTH) break;
-                float weight = FILTER_RWING[h_i] + interp * FILTER_DIFFS[h_i];
+                double weight = FILTER_RWING[h_i] + interp * FILTER_DIFFS[h_i];
                 dst_sample += src[s_i] * weight;
             }
